+++ NEWS.md
@@ -1,3 +1,8 @@
+# fabricQueryR 0.2.0
+
+* Added `fabric_livy_query()` to run queries against the 'Fabric Livy API',
+allowing remote execution of 'Spark' code.
+
 # fabricQueryR 0.1.1
 
 * Initial CRAN submission.